refactor: split long files for maintainability/clarity.
