+++ src/types/linear/comment.ts
@@ -1,0 +1,5 @@
+// Comment-related argument interfaces
+export interface CreateCommentArgs {
+  issueId: string;     // ID of the issue to comment on
+  body: string;        // Comment content
+}
